+++ sim/config.tcl
@@ -53,7 +53,6 @@ add_sources "../hdl" {
 	psi_common_par_tdm.vhd \
 	psi_common_tdm_par.vhd \
 	psi_common_tdm_par_cfg.vhd \
-	psi_common_tdm_par_fill.vhd \
 	psi_common_arb_priority.vhd \
 	psi_common_arb_round_robin.vhd \
 	psi_common_tdm_mux.vhd \
@@ -85,7 +84,7 @@ add_sources "../hdl" {
   psi_common_ser_par.vhd \
   psi_common_spi_master_cfg.vhd \
   psi_common_find_min_max.vhd \
-  psi_common_min_max_mean.vhd \
+  psi_common_min_max_sum.vhd \
 } -tag src
 
 # testbenches
@@ -108,7 +107,6 @@ add_sources "../testbench" {
 	psi_common_par_tdm_tb/psi_common_par_tdm_tb.vhd \
 	psi_common_tdm_par_tb/psi_common_tdm_par_tb.vhd \
 	psi_common_tdm_par_cfg_tb/psi_common_tdm_par_cfg_tb.vhd \
-	psi_common_tdm_par_fill_tb/psi_common_tdm_par_fill_tb.vhd \
 	psi_common_arb_priority_tb/psi_common_arb_priority_tb.vhd \
 	psi_common_arb_round_robin_tb/psi_common_arb_round_robin_tb.vhd \
 	psi_common_tdm_mux_tb/psi_common_tdm_mux_tb.vhd \
@@ -153,11 +151,11 @@ add_sources "../testbench" {
   psi_common_ser_par_tb/psi_common_ser_par_tb.vhd \
   psi_common_spi_master_cfg_tb/psi_common_spi_master_cfg_tb.vhd \
   psi_common_find_min_max_tb/psi_common_find_min_max_tb.vhd \
-  psi_common_min_max_mean_tb/psi_common_min_max_mean_tb.vhd \
+  psi_common_min_max_sum_tb/psi_common_min_max_sum_tb.vhd \
 } -tag tb
 
 #TB Runs
-create_tb_run "psi_common_min_max_mean_tb"
+create_tb_run "psi_common_min_max_sum_tb"
 tb_run_add_arguments \
   "-gclock_cycle_g=100 -gsigned_data_g=true -gdata_length_g=16 -gaccu_length_g=64" \
   "-gclock_cycle_g=10 -gsigned_data_g=false -gdata_length_g=8 -gaccu_length_g=16" \
@@ -315,9 +313,6 @@ create_tb_run "psi_common_tdm_par_tb"
 add_tb_run
 
 create_tb_run "psi_common_tdm_par_cfg_tb"
-add_tb_run
-
-create_tb_run "psi_common_tdm_par_fill_tb"
 add_tb_run
 
 create_tb_run "psi_common_arb_priority_tb"